third_party/WebKit/LayoutTests/accessibility/listbox-enabled-states-expected.txt


This tests that we report the correct enabled state on a listbox.


PASS axSelectElement.restrictions is "none"
PASS axSelectElement.childAtIndex(0).restrictions is "none"
PASS axSelectElement.childAtIndex(1).restrictions is "none"
PASS axSelectElement.childAtIndex(2).restrictions is "disabled"
PASS axSelectElement.childAtIndex(3).restrictions is "disabled"
PASS axSelectElement.childAtIndex(4).restrictions is "none"
PASS successfullyParsed is true

TEST COMPLETE
